moving rules to com.godaddy.sonar.ruby.metricfu.rules
